Fix older days to adapt IntCodeComputer to Long

+++ src/main/java/Day5.java
@@ -2,7 +2,7 @@ import java.util.Queue;
 
 public class Day5 {
 
-    public int[] callComputer(String program, Queue<Integer> userInputs, Queue<Integer> outputs) {
+    public long[] callComputer(String program, Queue<Long> userInputs, Queue<Long> outputs) {
         IntCodeComputer intCodeComputer = new IntCodeComputer(program);
         intCodeComputer.process(userInputs, outputs);
         return intCodeComputer.getProgram();
