+++ src/main/java/Day7.java
@@ -2,16 +2,16 @@ import java.util.*;
 
 public class Day7 {
 
-    public int maxAmplification(String program) {
-        int maxOutput = Integer.MIN_VALUE;
-        for (List<Integer> combination : allPermutations()) {
-            int secondInput = 0;
+    public long maxAmplification(String program) {
+        long maxOutput = Integer.MIN_VALUE;
+        for (List<Long> combination : allPermutations()) {
+            long secondInput = 0;
             for (int i = 0; i < combination.size(); i++) {
                 IntCodeComputer computer = new IntCodeComputer(program);
-                Queue<Integer> inputs = new LinkedList<>();
+                Queue<Long> inputs = new LinkedList<>();
                 inputs.add(combination.get(i));
                 inputs.add(secondInput);
-                Queue<Integer> outputs = new LinkedList<>();
+                Queue<Long> outputs = new LinkedList<>();
                 computer.process(inputs, outputs);
                 secondInput = outputs.poll();
             }
@@ -20,13 +20,13 @@ public class Day7 {
         return maxOutput;
     }
 
-    private List<List<Integer>> allPermutations() {
-        List<List<Integer>> permutations = new LinkedList<>();
-        evaluateAllPermutations(5, new Integer[]{0, 1, 2, 3, 4}, permutations);
+    private List<List<Long>> allPermutations() {
+        List<List<Long>> permutations = new LinkedList<>();
+        evaluateAllPermutations(5, new Long[]{0L, 1L, 2L, 3L, 4L}, permutations);
         return permutations;
     }
 
-    public void evaluateAllPermutations(int n, Integer[] elements, List<List<Integer>> combinations) {
+    public void evaluateAllPermutations(int n, Long[] elements, List<List<Long>> combinations) {
         if (n == 1) {
             combinations.add(new ArrayList<>(Arrays.asList(elements)));
         } else {
@@ -42,25 +42,25 @@ public class Day7 {
         }
     }
 
-    private void swap(Integer[] input, int a, int b) {
-        int tmp = input[a];
+    private void swap(Long[] input, int a, int b) {
+        long tmp = input[a];
         input[a] = input[b];
         input[b] = tmp;
     }
 
-    private List<List<Integer>> allFeedbackPermutations() {
-        List<List<Integer>> permutations = new LinkedList<>();
-        evaluateAllPermutations(5, new Integer[]{5, 6, 7, 8, 9}, permutations);
+    private List<List<Long>> allFeedbackPermutations() {
+        List<List<Long>> permutations = new LinkedList<>();
+        evaluateAllPermutations(5, new Long[]{5L, 6L, 7L, 8L, 9L}, permutations);
         return permutations;
     }
 
-    public int maxAmplificationInFeedbackLoop(String program) {
-        int maxOutput = Integer.MIN_VALUE;
-        for (List<Integer> combination : allFeedbackPermutations()) {
+    public long maxAmplificationInFeedbackLoop(String program) {
+        long maxOutput = Integer.MIN_VALUE;
+        for (List<Long> combination : allFeedbackPermutations()) {
 
             //create amplifier feedback loop
             Amplifier[] amplifiers = new Amplifier[5];
-            Queue<Integer>[] signals = new Queue[5];
+            Queue<Long>[] signals = new Queue[5];
             for (int i = 0; i < 5; i++) {
                 signals[i] = new LinkedList<>();
                 signals[i].offer(combination.get(i));
@@ -70,7 +70,7 @@ public class Day7 {
             }
 
             //add first signal to first amplifier
-            signals[0].offer(0);
+            signals[0].offer(0L);
 
             boolean computationCleared = false;
             while (!computationCleared) {
@@ -87,17 +87,17 @@ public class Day7 {
     private class Amplifier {
         private final String program;
         IntCodeComputer computer;
-        Queue<Integer> inputs;
-        Queue<Integer> outputs;
+        Queue<Long> inputs;
+        Queue<Long> outputs;
 
-        public Amplifier(String program, Queue<Integer> inputs, Queue<Integer> outputs) {
+        public Amplifier(String program, Queue<Long> inputs, Queue<Long> outputs) {
             this.program = program;
             computer = new IntCodeComputer(program);
             this.inputs = inputs;
             this.outputs = outputs;
         }
 
-        public Integer pollOutputs() {
+        public long pollOutputs() {
             return outputs.poll();
         }
